ai-server/requirements.txt
adafruit-io==2.7.0
tensorflow==2.12.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pymongo==4.3.3
PyAudio==0.2.13
nltk==3.8.1
python-dotenv==1.0.0
openai==0.27.6